feature: refactor position update methods to use 'plus' for vector calculations
